Alex feature reflect schema and add validation

+++ client/src/pages/job-dashboard-page/JobDashboard.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Button } from "react-bootstrap";
 import { Link } from "react-router-dom";
 import BootstrapTable from 'react-bootstrap-table-next';
-import cellEditFactory from 'react-bootstrap-table2-editor';
+import cellEditFactory, { Type } from 'react-bootstrap-table2-editor';
 import "./JobDashboard.css";
 
 const columns = [{
@@ -16,11 +16,11 @@ const columns = [{
   headerStyle: { color: '#347571'},
   style: {color: '#224e4b'}
 }, {
-  // dataField: 'location',
-  // text: 'Location',
-  // headerStyle: { color: '#347571'},
-  // style: {color: '#224e4b'}
-// }, {
+  dataField: 'location',
+  text: 'Location',
+  headerStyle: { color: '#347571'},
+  style: {color: '#224e4b'}
+}, {
   dataField: 'application',
   text: 'Application',
   headerStyle: { color: '#347571'},
@@ -28,13 +28,26 @@ const columns = [{
 }, {
   dataField: 'type',
   text: 'Type',
+  editor: {
+    type: Type.SELECT,
+    options: [
+      {
+        value: "INTERNSHIP",
+        label: "INTERNSHIP"
+      },
+      {
+        value: "FULLTIME",
+        label: "FULLTIME"
+      }
+    ]
+  },
   headerStyle: { color: '#347571'},
   style: {color: '#224e4b'}
-// }, {
-  // dataField: 'description',
-  // text: 'Description',
-  // headerStyle: { color: '#347571'},
-  // style: {color: '#224e4b'}
+}, {
+  dataField: 'description',
+  text: 'Description',
+  headerStyle: { color: '#347571'},
+  style: {color: '#224e4b'}
 }, {
   editable: false,
   formatter: (content, row) => {
@@ -47,17 +60,19 @@ const columns = [{
 }];
 
 const deleteJob = (jobId) => {
-  const token = localStorage.getItem("token");
-  fetch(`api/jobs/${jobId}`, {
-    method: 'DELETE',
-    headers: {
-      "Content-Type": "application/json",
-      "Authorization": `Bearer ${token}`
-    }
-  }).then(onSubmitSuccess().catch(err => console.err(err)))
+  if (window.confirm("Are you sure you want to delete this job?")) {
+    const token = localStorage.getItem("token");
+    fetch(`api/jobs/${jobId}`, {
+      method: 'DELETE',
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
+    }).then(onSubmitSuccess().catch(err => console.err(err)))
+  }
 }
 
-const onSubmitSuccess= () => {
+const onSubmitSuccess = () => {
   window.location.href = "/job-dashboard"
 }
 
@@ -96,8 +111,10 @@ function JobDashboard() {
         id: job._id,
         title: job.jobTitle, 
         company: job.jobCompany,
+        location: job.jobLocation,
         application: job.appLink,
-        type: job.jobType
+        type: job.jobType,
+        description: job.jobDescription
       }
 
       dataArray.push(each)
@@ -110,8 +127,10 @@ function JobDashboard() {
     let values = {
       jobTitle: row.title,
       jobCompany: row.company,
+      jobLocation: row.location,
       appLink: row.application,
-      jobType: row.type
+      jobType: row.type,
+      jobDescription: row.description
     }
 
     // PUT to DB
